refactor(snapshots): made DirManifestBuilder public
No other changes.
